+++ tools/keyShortcuts/tool.js
@@ -1,6 +1,8 @@
 (() => {
   class KeyShortcutsTool extends LiChessTools.Tools.ToolBase {
 
+    dependencies = ['EmitRedraw','ExplorerPractice','AdditionalGlyphs','SearchMovesCommand'];
+
     preferences = [
       {
         name: 'keyShortcuts',
@@ -222,6 +224,7 @@
 
     bindKeysForAnalysis = () => {
       const lt = this.lichessTools;
+      const $ = lt.$;
       const analysis = lt.lichess.analysis;
       lt.unbindKeyHandler('i');
       lt.unbindKeyHandler('m');
@@ -231,6 +234,8 @@
       lt.unbindKeyHandler('alt+m', true);
       lt.unbindKeyHandler('alt+b', true);
       lt.unbindKeyHandler('alt+g', true);
+      lt.unbindKeyHandler('o', true);
+      lt.unbindKeyHandler('alt+o', true);
 
       lt.unbindKeyHandler('.', true);
       lt.unbindKeyHandler('ctrl+.', true);
@@ -265,6 +270,10 @@
         lt.bindKeyHandler('alt+m', () => lt.jumpToGlyphSymbols('?', true));
         lt.bindKeyHandler('alt+b', () => lt.jumpToGlyphSymbols('??', true));
         lt.bindKeyHandler('alt+g', () => lt.jumpToGlyphSymbols(['!', '!?', '!!', lt.icon.WhiteStar], true));
+        if (lt.tools.AdditionalGlyphsTool?.options?.slow && !$('span.lichessTools-obsSetup').length) {
+          lt.bindKeyHandler('o', () => this.jumpToSlowMoves(false));
+          lt.bindKeyHandler('alt+o', () => this.jumpToSlowMoves(true));
+        }
 
         lt.bindKeyHandler('.', () => this.prepareMove('pgn'));
         lt.bindKeyHandler('ctrl+.', () => this.prepareMove('ceval'));
@@ -294,6 +303,32 @@
           }
         }
       }
+    };
+
+    jumpToSlowMoves = (black) => {
+      const lt = this.lichessTools;
+      const analysis = lt.lichess?.analysis;
+      if (!analysis) return;
+      const glyphsTool = lt.tools.AdditionalGlyphsTool;
+      if (!glyphsTool.options.slow) return;
+
+      const nodes = [];
+      let path = '';
+      for (const node of analysis.mainline) {
+        path+=node.id;
+        if (node.isSlow === undefined) {
+          glyphsTool.processSlow(node);
+        }
+        if (node.isSlow && node.ply % 2 == !black) {
+          nodes.push(path);
+        }
+      }
+      if (!nodes.length) return;
+
+      let index = nodes.indexOf(analysis.path);
+      index = (index<0 ? 0 : index+1) % nodes.length;
+      analysis.userJumpIfCan(nodes[index]);
+      lt.analysisRedraw();
     };
 
     handleEditorAction = (index) => {